Crate math_api_experiments

Modules

v1
A partial implementation of some vector maths using the same type of tricks that gecko currently uses to represent units. The idea is to compare with the approach in v2.rs which is closer to what euclid does.

v2
A partial implementation of some vector maths using the same type of tricks that euclid currently uses to represent units. The difference is that euclid currently can't express a matrix that transform points from a certain unit into points with another units, for example a projection matrix that would take WorldPoints and create ScreenPoints.