Module math_api_experiments::v1

A partial implementation of some vector maths using the same type of tricks that gecko currently uses to represent units. The idea is to compare with the approach in v2.rs which is closer to what euclid does.

Structs

Matrix4x4
A 4 by 4 matrix that can represent a transformation from a space to another.

Point3D
A 3d vector.

Screen

Untyped

World