Module math_api_experiments::v2

A partial implementation of some vector maths using the same type of tricks that euclid currently uses to represent units. The difference is that euclid currently can't express a matrix that transform points from a certain unit into points with another units, for example a projection matrix that would take WorldPoints and create ScreenPoints.

Structs

Matrix4x4
A 4 by 4 matrix.

TransformComponent
A scalar that is used to transform values from a space to another.

Typed
A scalar that is expressed in a certain space.

Traits

CustomOne
num::One requires Self to implement Mul which we can't implement for TransformComponent, so we need a custom trait that provides the same functionality.